+++ src/components/App.js
@@ -1,6 +1,5 @@
 import React from 'react'
 
-import AppBar from './AppBar'
 import ReservationForm from './ReservationForm'
 import ReservationPriceTable from './ReservationPriceTable'
 import ErrorMessage from './ErrorMessage'
@@ -8,7 +7,6 @@ import ErrorMessage from './ErrorMessage'
 export default function App() {
   return (
     <div>
-      <AppBar />
       <ReservationForm />
       <ReservationPriceTable />
       <ErrorMessage />
